Update website layout and content
The layout and content of the website have been completely revamped. The previous version was more focused on jokes and memes, however, the new version is turned into a personal portfolio. This allows the user to showcase their work and provide information about themselves. The CSS file has been updated to change the look and feel of the website to be cleaner and more professional. All joke-related content has been removed from the HTML and Javascript files and replaced with standard portfolio sections such as 'About', 'Projects', and 'Contact'.